PostRepositoryを抽象クラスで実装するのをやめた
