+++ src/repositories/PostsRepository.test.ts
@@ -3,7 +3,7 @@ import { load } from 'cheerio';
 import { instance, mock, when } from 'ts-mockito';
 import { profile } from '@/config/profile';
 import { Post } from '@/entities';
-import { PostsRepositoryImpl } from './PostsRepositoryImpl';
+import { PostsRepository } from './PostsRepository';
 import type cpx from 'cpx';
 
 vi.mock('cpx', async () => {
@@ -15,16 +15,16 @@ vi.mock('cpx', async () => {
     };
 });
 
-describe('PostsRepositoryImpl', () => {
+describe('PostsRepository', () => {
     beforeAll(() => {
-        Object.defineProperty(PostsRepositoryImpl, 'postsDirectory', {
+        Object.defineProperty(PostsRepository, 'postsDirectory', {
             value: () => join(process.cwd(), 'test/content/posts'),
         });
     });
 
     describe('getAllPosts', () => {
         it('returns all posts', () => {
-            const repo = new PostsRepositoryImpl();
+            const repo = new PostsRepository();
             const result = repo.getAllPosts();
 
             expect(result.length).toBe(1);
@@ -45,7 +45,7 @@ describe('PostsRepositoryImpl', () => {
 
     describe('getPostBySlug', () => {
         it('returns post by slug', () => {
-            const repo = new PostsRepositoryImpl();
+            const repo = new PostsRepository();
             const result = repo.getPostBySlug('test-post');
             expect(result).not.toBeNull();
 
@@ -60,7 +60,7 @@ describe('PostsRepositoryImpl', () => {
         });
 
         it('returns null if slug does not exists', () => {
-            const repo = new PostsRepositoryImpl();
+            const repo = new PostsRepository();
             const result = repo.getPostBySlug('not-exist-slug');
             expect(result).toBeNull();
         });
@@ -68,7 +68,7 @@ describe('PostsRepositoryImpl', () => {
 
     describe('getPostsByTag', () => {
         it('returns posts by tag', () => {
-            const repo = new PostsRepositoryImpl();
+            const repo = new PostsRepository();
             const mockPosts: Post[] = [...Array(3)].map(() => mock<Post>());
             when(mockPosts[0].tags).thenReturn([
                 'TypeScript',
@@ -85,7 +85,7 @@ describe('PostsRepositoryImpl', () => {
         });
 
         it('ignores Upper or Lower', () => {
-            const repo = new PostsRepositoryImpl();
+            const repo = new PostsRepository();
             const mockPosts: Post[] = [mock<Post>()];
             when(mockPosts[0].tags).thenReturn(['TypeScript']);
             Object.defineProperty(repo, 'getAllPosts', {
